cs - Moved 'Reset all' button to the my schedule's table
